backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pydantic>=2.10.0
pydantic-settings>=2.0.0
pydantic-ai>=0.0.14
asyncpg==0.29.0
httpx>=0.27.2
beautifulsoup4==4.12.3
alembic==1.13.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
lxml==5.1.0
apscheduler==3.10.4
openai>=1.12.0
feedparser>=6.0.10
jupyterlab
pandas
matplotlib
sqlalchemy
psycopg2-binary
